organized files, created beginning character folder and files.
