+++ GadzOOks/GadzOOks/Character/PlayerCharacter.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GadzOOks.Character
+{
+    internal class PlayerCharacter
+    {
+    }
+}
